refactor: domain하위 values, enums 패키지 추가 및 이동
- Level을 enums 하위로 이동
- LogData를 values 하위로 이동
- Test코드를 해당 패키지 구조에 맞게 변경
