+++ logbat/src/test/java/info/logbat/log/domain/LogTest.java
@@ -3,6 +3,7 @@ package info.logbat.log.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import info.logbat.log.domain.enums.Level;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
